removed unused files and directories + fixed the norme in libft file, added extra memory file for gc management
