+++ libft/ft_lstremove.c
@@ -1,5 +1,16 @@
-# include "./libft.h"
-# include <stdio.h>
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_lstremove.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sboukiou <sboukiou@student.1337.ma>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/28 02:13:52 by sboukiou          #+#    #+#             */
+/*   Updated: 2025/07/28 02:14:00 by sboukiou         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "./libft.h"
 
 /**
  * ft_ltsremove - removes a node from a linked-list
